+++ core/src/com/mygdx/tdt4240/utils/Constants.kt
@@ -5,8 +5,8 @@ import java.awt.Color
 
 object Constants {
     const val GAME_TITLE = "Samf Survivor"
-    val GAME_WIDTH = Gdx.graphics.width.toFloat()
-    val GAME_HEIGHT = Gdx.graphics.height.toFloat()
+    val GAME_WIDTH = Gdx.graphics.width
+    val GAME_HEIGHT = Gdx.graphics.height
     const val SAMF_RED= "A03033"
     const val SAMF_RED_DARK="5B0E10"
     const val BLACK="190304"
